feat(nvim): remove bufferline et al
I waste too much time trying to organize the tabline of open buffers...
for no particular reason.

Because of it, I also had the tendency of trying to navigate with :bprev
and :bnext, and it is almost never a good experience.

I'm gonna experiment with bufferline and the shortcuts for it, as well
as the bufdelete plugin and related keybindings.

I can still close all buffers with leader+bad, close a particular buffer
with leader+q, and close all buffers but the current one with
leader+bsd.

This last one I don't think I'll still use much, to be honest, it was
mostly to keep the bufferline clean, and I don't have a bufferline
anymore... so... let's see.