fix(nvim): gopls.tidy fix
Apparently the copilot LSP listens to every LSP request, and if it
doesn't handle it, it just errors, which ends up canceling everything (I
guess).

Inspired more or less on what ray-x/go.nvim does, which seems to work better:
we send the request only to gopls - which makes sense, as I think only
gopls can reply to `gopls.tidy`.

Anyway, this fixes more things than it seems to break, so yolo.